python_server/pyproject.toml
[project]
name = "python-server"
version = "0.1.0"
description = "Crypto scraping server using Playwright + Beanie"
authors = [
    { name = "elchanan" }
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "playwright>=1.52.0,<2.0.0",
    "beanie>=1.24.0,<2.0.0",
    "motor>=3.4.0,<4.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "psutil (>=7.0.0,<8.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "redis (>=6.2.0,<7.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
schedule = "^1.2.2"
[tool.poetry]
package-mode = false